requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
alembic>=1.13.1
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
textblob>=0.17.1
openai>=1.8.0
pydantic>=2.5.3
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.1.2
pytest>=7.4.4
httpx>=0.26.0  # for testing
python-jose[cryptography]>=3.3.0  # for JWT

